+++ common/src/main/java/dev/ripio/cobbleloots/network/CobblelootsClientNetworkSender.java
@@ -1,0 +1,10 @@
+package dev.ripio.cobbleloots.network;
+
+import dev.architectury.injectables.annotations.ExpectPlatform;
+
+public class CobblelootsClientNetworkSender {
+    @ExpectPlatform
+    public static void sendLootBallUpdate(CobblelootsLootBallUpdatePayload payload) {
+        throw new AssertionError();
+    }
+}