+++ fabric/src/main/java/dev/ripio/cobbleloots/network/fabric/CobblelootsNetworkSenderImpl.java
@@ -1,0 +1,11 @@
+package dev.ripio.cobbleloots.network.fabric;
+
+import dev.ripio.cobbleloots.network.CobblelootsLootBallOpenScreenPayload;
+import net.fabricmc.fabric.api.networking.v1.ServerPlayNetworking;
+import net.minecraft.server.level.ServerPlayer;
+
+public class CobblelootsNetworkSenderImpl {
+    public static void sendLootBallOpenScreen(ServerPlayer player, CobblelootsLootBallOpenScreenPayload payload) {
+        ServerPlayNetworking.send(player, payload);
+    }
+}